Login with People OpenId
You can now login via OpenId using directed identity with your eXtension ID. This creates an authmap record and an associated learner record with name and email address populated from people.
Still a few loose ends that need to be tied up. Also incuded, is a fix for the edit comment link for non-logged in learners.