Login Page Added and Devise Fixes
Added database_authenticatable to learner so we can get our login page routes and so that the redirects will work correctly with devise. Also included is the new login page with the links to login using both twitter and eXtension OpenId.

Also included were some fixes for links and the devise adjustments fixed an issue that was causing an app error when the authentication failed for a service, now the learner is redirected back to the login page if they cannot login to a service. Included in this is a conditional for comments on a event page that displays a link to sign in instead of a post button if the learner is not logged in since we have a authentication requirement for comments.