move ported finds to model, fix other scope
takes advantage of the scoping in Active Relation, follows active relation practice recommendations for the controller.  Note:  (limit =3) in controller is syntactic sugar to give a better visible indicator of the value being used in the lambda.  i.e.  Event.upcoming(5) is less clear than Event.upcoming(limit = 5) - the (limit =5) evaluates to 5 as well, and gives an indicator of what '5' is when triggering the scope.

through_this_week needed to be wrapped in a lambda to ensure that the times were evaluated at runtime.