Learner Portfolio Page
Event information is now populated on this page. This page is waiting on code to be written for learner profile information in order to populate Current Learning Items, Learning Overview, Learner Bio, and views and controllers need to be created for the view links on this page to link to. A decision on limits of items on this page needs to be made as well.