Addresses Issue #4 Added Google OpenID Auth and Added Back Twitter Auth
Along with this is a feature that will search for a learner with the same email address as that returned from the authentication service (if applicable) and use that learner for the authmap if the authmap did not already exist for that service. This will help to merge accounts that might otherwise be separate for the same learner. Still waiting on facebook authentication to close this one out.