refactor link_to_learner and get_avatar
get_avatar was confusing and difficult to follow for me - refactored into link_to_learner and link_to_learner_avatar and avatar_for_learner - which share an api, and describe a bit better the intent of the method call. link_path can also be overridden - and thus useable in a conference context.